requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/base.txt requirements/base.in
#
attrs==19.2.0             # via environ-config
beautifulsoup4==4.7.1
cachetools==3.1.1         # via google-auth
certifi==2019.6.16        # via httpx, requests
chardet==3.0.4            # via httpx, requests
click==7.0
environ-config==19.1.0
google-api-core==1.13.0   # via google-cloud-core
google-auth==1.6.3        # via google-api-core
google-cloud-bigquery==1.15.0
google-cloud-core==1.0.2  # via google-cloud-bigquery
google-resumable-media==0.3.2  # via google-cloud-bigquery
googleapis-common-protos==1.6.0  # via google-api-core
h11==0.8.1                # via httpx
h2==3.1.1                 # via httpx
hpack==3.0.0              # via h2
hstspreload==2019.9.30    # via httpx
httpx==0.7.4
hyperframe==5.2.0         # via h2
idna==2.8                 # via httpx, requests
lxml==4.3.4
pendulum==2.0.4
protobuf==3.8.0           # via google-api-core, google-cloud-bigquery, googleapis-common-protos
pyasn1-modules==0.2.5     # via google-auth
pyasn1==0.4.5             # via pyasn1-modules, rsa
python-dateutil==2.8.0    # via pendulum
pytz==2019.1              # via google-api-core
pytzdata==2019.1          # via pendulum
requests==2.22.0          # via google-api-core
rfc3986==1.3.2            # via httpx
rsa==4.0                  # via google-auth
six==1.12.0               # via google-api-core, google-auth, google-resumable-media, protobuf, python-dateutil
soupsieve==1.9.2          # via beautifulsoup4
urllib3==1.25.3           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.4.0        # via google-api-core, protobuf
